Overwrite in determining file directory

+++ python/scan.py
@@ -20,6 +20,7 @@ import prescan
 from utils import fileutils
 
 import copy
+from utils import tsvutils
 
 # class to organize and run a complete scan
 class Scan:
@@ -28,7 +29,8 @@ class Scan:
                  masses: 'Masses',
                  modelname,
                  decay,
-                 maxwidth):
+                 maxwidth, 
+                 overwrite=False):
 
         # store model name
         self.modelname = modelname
@@ -58,14 +60,14 @@ class Scan:
         self.outdir = fileutils.scanDir(modelname=modelname,decay=decay,masses=masses)
 
         # remove previous directory if set to overwrite
-        if os.path.exists(self.outdir):
+        if os.path.exists(self.outdir) and overwrite:
+            # remove directory
             shutil.rmtree(self.outdir)
 
-        # make working directory
-        os.makedirs(self.outdir)
-
-        # directory to store all of the output files
-        os.makedirs(self.outdir+"/files")
+        # check if directory exists, if not make it
+        if not os.path.exists(self.outdir):
+            os.makedirs(self.outdir)
+            os.makedirs(self.outdir+"/files")
 
         # create summary file
         self.summaryname = self.outdir+"scansummary_"+self.modelname+"_"+self.decay+"_"+str(self.masses)+".txt"
@@ -130,6 +132,8 @@ class Scan:
                                    masses=self.masses,
                                    modelname=self.modelname,
                                    decay=self.decay)
+        
+        #print(self.prescanparser.getArrays())
 
         # if the prescan ranges are more than 5% away from
         # the boundaries, change the boundaries to restrict
@@ -160,6 +164,7 @@ class Scan:
             # print min and max to screen after prescan
             self.params.printMinMax(par)
 
+        
         # get scan density
         density = nprescan / self.params.volume()
 
@@ -209,15 +214,33 @@ class Scan:
         # run prescan
         self.runPrescan(npoints=npoints,
                         use_multiprocessing=use_multiprocessing)
-
+        
         # move into the working directory for scans
         os.chdir(self.outdir)
 
-        #Make copies of the params object
-        s1_params = copy.deepcopy(self.params)
+        all_scanners = []
 
-        scanner1 = Scanner(npoints=npoints,
-                            params=s1_params,
+        #Iterate through the prescan to determine if multiple scanners are needed
+        for par in self.params.parnames:
+          
+            if self.prescanparser.isBimodal(par):
+                print("Bimodal - Breaking into 2 scanners")
+                
+                params_1 = copy.deepcopy(self.params)
+                params_2 = copy.deepcopy(self.params)
+                            
+                p_min = params_1.low(par)
+                p_max = params_1.high(par)
+                p_mid = (p_min + p_max) / 2.0
+                
+                params_1.setMin(par, p_min)
+                params_1.setMax(par, p_mid)
+                
+                params_2.setMin(par, p_mid)
+                params_2.setMax(par, p_max)
+
+                firstScanner = Scanner(npoints=npoints,
+                            params=params_1,
                             decay=self.decay,
                             maxwidth=self.maxwidth,
                             optPoint=self.optPoint,
@@ -225,94 +248,82 @@ class Scan:
                             summaryname=self.summaryname,
                             zoom=zoom,
                             outdir=self.outdir,
-                            label="s-1")
+                            label=f'{par}-1')
+                
+                secondScanner = Scanner(npoints=npoints,
+                            params=params_2,
+                            decay=self.decay,
+                            maxwidth=self.maxwidth,
+                            optPoint=self.optPoint,
+                            detailsname=self.detailsname,
+                            summaryname=self.summaryname,
+                            zoom=zoom,
+                            outdir=self.outdir,
+                            label=f'{par}-2')
+                
+                all_scanners.extend([firstScanner, secondScanner])
+            else:
+                #Make copies of the params object
+                s1_params = copy.deepcopy(self.params)
+
+                scanner1 = Scanner(npoints=npoints,
+                                    params=s1_params,
+                                    decay=self.decay,
+                                    maxwidth=self.maxwidth,
+                                    optPoint=self.optPoint,
+                                    detailsname=self.detailsname,
+                                    summaryname=self.summaryname,
+                                    zoom=zoom,
+                                    outdir=self.outdir,
+                                    label=f'{par}')
+                all_scanners.append(scanner1)
         
-        all_scanners = [scanner1]
+        #curr_opt = all_scanners[0].scanparser.getMaxPoint()
 
-        #param_dict = {scanner1.params.parnames: scanner1}
-
-        for par in scanner1.params.parnames:
-
-            if scanner1.scanparser.isBimodal(par):
-
-                params_1 = copy.deepcopy(self.params)
-                params_2 = copy.deepcopy(self.params)
-
-                print("Now printing params 1 ...")
-                for par in params_1.parnames:
-
-                    p_min = params_1.low(par)
-                    p_max = params_1.high(par)
-                    p_mid = params_1.getMidPoint(p_min, p_max)
-
-                    params_1.setMin(par, p_min)
-                    params_1.setMax(par, p_mid)
-
-                    params_1.printMinMax(par)
-
-                print("Now printing params 2 ...")
-                for par in params_2.parnames:
-
-                    p_min = params_2.low(par)
-                    p_max = params_2.high(par)
-                    p_mid = (p_min + p_max) / 2
-
-                    params_2.setMin(par, p_mid)
-                    params_2.setMax(par, p_max)
-
-                    params_2.printMinMax(par)
-
-            # TODO: Need to find an optPoint for each scanner range
-            firstscanner = Scanner(npoints=npoints,
-                                params=params_1,
-                                decay=self.decay,
-                                maxwidth=self.maxwidth,
-                                optPoint=self.optPoint,
-                                detailsname=self.detailsname,
-                                summaryname=self.summaryname,
-                                zoom=zoom,
-                                outdir=self.outdir,
-                                label="test-1")
-
-            secondScanner = Scanner(npoints=npoints,
-                                params=params_2,
-                                decay=self.decay,
-                                maxwidth=self.maxwidth,
-                                optPoint=self.optPoint,
-                                detailsname=self.detailsname,
-                                summaryname=self.summaryname,
-                                zoom=zoom,
-                                outdir=self.outdir,
-                                label="test-2")
-            
-            all_scanners.extend([firstscanner, secondScanner])
-
-        count = 0
-        all_optpoints = []
-        # run multiple scan iterations
         for iter in range(niter):
 
-            # run scanner
-            myscanner.run(iter,use_multiprocessing)
+            for scans in all_scanners:
 
-            #check_all(myscanner)
+                scans.run(iter, use_multiprocessing)
 
-            secondScanner.run(iter,use_multiprocessing)
+                '''thresh = scans.scanparser.getMaxPoint() * 0.05
 
+                diff = abs(scans.scanparser.getMaxPoint() - curr_opt)
+
+                if diff <= thresh:
+                    break
+                
+                curr_opt = scans.scanparser.getMaxPoint()'''
+
+        
             ##### TODO: Add early stopping conditions
             #all_optpoints.append(myscanner.optPoint.getVal(params_1[varname=]))
-            '''
-            if iter > 0:
-                p1 = all_optpoints[iter - 1]
-                p2 = all_optpoints[iter]
-
-                thresh = p1 * 0.05
-
-                if abs(p2-p1) <= thresh:
-                    break
-            '''
+           
             ##### TODO: Add functionality to concatenate all outputs into a single large output
+
+        if bimodal:
+
+            all_tsvs = []
+
+            for file in self.outdir:
+
+                if ".tsv" in file:
+                    all_tsvs.append(file)
             
+            inp = ""
+            outp = ""
+            combine = False
+            for tsv in all_tsvs:
+                
+                if "MultScanner-1" in tsv:
+                    inp = tsv
+                
+                if "MultScanner-2" in tsv:
+                    outp = tsv
+                    combine = True
+
+                if combine:
+                    tsvutils.saveTSVOutput(outp, inp)
 
         # get total scan time
         scanend = time.time()
@@ -326,12 +337,8 @@ class Scan:
         details = open(self.detailsname,"a")
         details.write("Scan took "+str(datetime.timedelta(seconds=int(scantime)))+" (hh:mm:ss)")
         details.close()
-
+        
         return
-    
-'''def check_all(scan):
-    
-    for par, names in '''
 
 def updateParams(all_params):
 
@@ -581,6 +588,7 @@ if __name__ == "__main__":
     argparser.add_argument("-r", "--parameter_rate", default=0.05, type=float, help="Rate at which parameter range should shrink")
     argparser.add_argument("-g", "--density_growth", default=0.2, type=float, help="Rate at which point density should grow")
     argparser.add_argument("-m", "--multiprocessing", action="store_true", help="Whether multiprocessing should be used")
+    argparser.add_argument("-o", "--overwrite", action="store_true", help="Overwrite previous prescan")
     args = argparser.parse_args()
 
     # create masses object
@@ -594,7 +602,8 @@ if __name__ == "__main__":
     myScan = Scan(masses=masses,
                   modelname=args.model,
                   decay=args.decay,
-                  maxwidth=args.maxwidth)
+                  maxwidth=args.maxwidth,
+                  overwrite=args.overwrite)
     
     # run scan using scan object
     myScan.runScan(npoints=args.npoints,
